pyproject.toml
[project]
name = "poetry1"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey",email = "andrey.kin.1984@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "setuptools (>=80.3.1,<81.0.0)",
    "make (>=0.1.6.post2,<0.2.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=6.0.0,<7.0.0)",
]

[tool.poetry]
packages = [{include = "poetry1", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
